move find_free_port to iris.cluster.net to fix inverted dependency
kubectl.py was importing from platform/base.py, creating a dependency
from the k8s layer back into the platform layer. Move find_free_port
to a shared net module and update all call sites.
