+++ lib/iris/tests/e2e/_docker_cluster.py
@@ -17,7 +17,7 @@ from iris.client import IrisClient
 from iris.cluster.controller.controller import Controller, ControllerConfig
 from iris.cluster.controller.worker_provider import RpcWorkerStubFactory, WorkerProvider
 from iris.cluster.local_cluster import LocalCluster
-from iris.cluster.platform.base import find_free_port
+from iris.cluster.net import find_free_port
 from iris.cluster.runtime.docker import DockerRuntime
 from iris.cluster.types import Entrypoint, EnvironmentSpec, JobName, ResourceSpec
 from iris.cluster.bundle import BundleStore
